added race data
