+++ package.json
@@ -13,6 +13,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
+    "react-select": "^5.7.3",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0"
   },
