+++ public/races/2023/ajna-may03.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["IllIllI-bot"],
+	"A": ["MrsHudson", "Hound", "LightChaser"],
+	"B": ["c4lanky", "The_Madaladinator"],
+	"C": ["znBotty", "iceBearRobot", "Gas-Optimzeer", "sol2vuln"],
+	"prize": {
+        "winner": 1381.61,
+        "A": 1062.78,
+        "B": 215.03
+    }    
+}
