+++ public/races/2023/basin-jul03.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["0x6980"],
+	"A": ["zz_plus_ultra"],
+	"B": ["MrsHudson", "sol2vuln", "c4lanky"],
+	"C": ["The_Madaladinator", "LightChaser", "Hound", "znBotty", "Gas-Optimzeer", "Decurity_AI_v0"],
+	"prize": {
+        "winner": 1075.79,
+        "A": 827.53,
+        "B": 203.06
+    }    
+}
