+++ public/races/2023/chainlink-may26.json
@@ -1,0 +1,11 @@
+{
+	"winners": ["IllIllI-bot", "LightChaser", "zz_plus_ultra"],
+	"A": ["The_Madaladinator", "MrsHudson", "c4lanky", "znBotty"],
+	"B": ["Hound"],
+	"C": ["Gas-Optimzeer"],
+	"prize": {
+        "winner": 802.02,
+        "A": 616.94,
+        "B": 126.17
+    }    
+}
