+++ public/races/2023/eigenlayer-apr27.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["IllIllI-bot"],
+	"A": ["znBotty"],
+	"B": ["MrsHudson", "zz_plus_ultra", "LightChaser", "Hound", "The_Madaladinator"],
+	"C": ["4lpine", "iceBearRobot", "c4lanky", "Narya", "sol2vuln", "SolidityScanTeam"],
+	"prize": {
+        "winner": 3054.66,
+        "A": 2349.74,
+        "B": 419.12
+    }    
+}
