+++ public/races/2023/llama-jun06.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["IllIllI-bot"],
+	"A": ["MrsHudson", "Hound", "Gas-Optimzeer", "c4lanky", "zz_plus_ultra", "The_Madaladinator"],
+	"B": ["znBotty", "LightChaser"],
+	"C": ["0x6980", "sol2vuln", "Decurity_AI_v0"],
+	"prize": {
+		"winner": 0,
+		"A": 0,
+		"B": 0
+	}
+}
