+++ public/races/2023/lukso-jun30.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["IllIllI-bot"],
+	"A": ["zz_plus_ultra", "MrsHudson", "The_Madaladinator", "LightChaser", "Hound", "znBotty"],
+	"B": ["c4lanky", "Gas-Optimzeer"],
+    "C": [],
+	"prize": {
+        "winner": 1082.00,
+        "A": 832.31,
+        "B": 130.82
+    }    
+}
