+++ public/races/2023/lybra-jun23.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["IllIllI-bot"],
+	"A": ["LightChaser", "znBotty", "MrsHudson"],
+	"B": ["Hound", "zz_plus_ultra", "c4lanky", "The_Madaladinator"],
+	"C": ["Gas-Optimzeer", "Regast"],
+	"prize": {
+        "winner": 988.86,
+        "A": 760.66,
+        "B": 119.79
+    }    
+}
