+++ public/races/2023/maia-may30.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["c4lanky"],
+	"A": ["IllIllI-bot", "The_Madaladinator"],
+	"B": ["Hound", "LightChaser"],
+	"C": ["MrsHudson", "Gas-Optimzeer", "znBotty", "sol2vuln"],
+	"prize": {
+        "winner": 2645.37,
+        "A": 3438.99,
+        "B": 635.13
+    }    
+}
