+++ public/races/2023/nounsdao-jul03.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["znBotty"],
+	"A": ["LightChaser", "Hound", "0x6980", "zz_plus_ultra", "Gas-Optimzeer"],
+	"B": ["MrsHudson"],
+	"C": ["The_Madaladinator", "c4lanky"],
+	"prize": {
+        "winner": 1270.74,
+        "A": 977.49,
+        "B": 179.31
+    }    
+}
