+++ public/races/2023/stader-jun02.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["IllIllI-bot"],
+	"A": ["c4lanky", "LightChaser", "The_Madaladinator", "znBotty"],
+	"B": ["Hound", "zz_plus_ultra", "MrsHudson"],
+	"C": ["0x6980"],
+	"prize": {
+		"winner": 0,
+		"A": 0,
+		"B": 0
+	}
+}
