+++ public/races/2023/tapioca-jul05.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["IllIllI-bot"],
+	"A": ["The_Madaladinator", "MrsHudson", "Hound", "znBotty"],
+	"B": ["LightChaser", "c4lanky", "zz_plus_ultra", "sol2vuln", "graphium-bot"],
+	"C": ["Gas-Optimzeer"],
+	"prize": {
+        "winner": 4335.53,
+        "A": 3335.02,
+        "B": 422.73
+    }    
+}
