+++ public/races/2023/venus-may08.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["IllIllI-bot"],
+	"A": ["znBotty"],
+	"B": ["LightChaser", "Hound", "zz_plus_ultra", "MrsHudson", "c4lanky", "The_Madaladinator"],
+	"C": ["Gas-Optimzeer", "Narya", "pyaduit", "sol2vuln"],
+	"prize": {
+        "winner": 3004.03,
+        "A": 2310.80,
+        "B": 364.19
+    }    
+}
