+++ public/races/2023/⁠juicebox-may18.json
@@ -1,0 +1,11 @@
+{
+	"winner": ["IllIllI-bot"],
+	"A": ["Gas-Optimzeer", "znBotty", "MrsHudson", "c4lanky", "zz_plus_ultra"],
+	"B": ["Hound"],
+	"C": ["iceBearRobot", "The_Madaladinator", "LightChaser", "sol2vuln"],
+	"prize": {
+        "winner": 401.02,
+        "A": 308.48,
+        "B": 56.59
+    }    
+}
