+++ src/App.css
@@ -1,21 +1,35 @@
 body {
   background-color: #222;
+  font-family: "Roboto", Arial, sans-serif;
+  margin: 0;
+  padding: 0;
+  
 }
 
 .container {
-  max-width: 400px;
+  max-width: 800px;
   margin: 0 auto;
   padding: 40px;
-  text-align: center;
-  background-color: #222;
+  background-color: #333;
   color: #fff;
+  border-radius: 12px;
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.3);
+  box-sizing: border-box;
+}
+
+@media (max-width: 480px) {
+  .container {
+    padding: 20px;
+  }
 }
 
 .title {
-  font-size: 36px;
+  font-family: "Roboto", Arial, sans-serif;
+  font-size: 32px;
   font-weight: bold;
-  margin-bottom: 20px;
+  margin-bottom: 30px;
   color: #ffcc00;
+  text-align: center;
 }
 
 .search-container {
@@ -26,9 +40,30 @@ body {
   width: 100%;
   padding: 10px;
   font-size: 16px;
-  background-color: #333;
+  background-color: #444;
   border: none;
+  border-radius: 4px;
   color: #fff;
+}
+
+.tab-container {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 20px;
+}
+
+.tab {
+  padding: 10px 20px;
+  cursor: pointer;
+  background-color: #555;
+  color: #fff;
+  border-radius: 4px 4px 0 0;
+  margin-right: 10px;
+  transition: background-color 0.3s ease;
+}
+
+.tab.active {
+  background-color: #222;
 }
 
 .results-container {
@@ -37,10 +72,10 @@ body {
 
 .result-item {
   margin-bottom: 20px;
-  background-color: #444;
-  padding: 10px;
+  background-color: #555;
+  padding: 20px;
   border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
 }
 
 .submitter {
